fix: Clean up code after PR-Review
