+++ src/components/stickable/list/OpenCollapsibleId.ts
