+++ src/i18n/list/list.en.json
@@ -9,15 +9,15 @@
     "status": {
       "personal": {
         "title": "Personal",
-        "description": "This is a personal trick that you added to your trick-list. No-one else will be able to see this trick in their trick list."
+        "description": "This is a personal trick that you added to your trick list. No-one else will be able to see this trick in their trick list."
       },
       "archived": {
         "title": "Archived",
-        "description": "This trick used to be part of the official trick-list but got removed at some point."
+        "description": "This trick used to be part of the official trick list but got removed at some point."
       },
       "new": {
         "title": "New",
-        "description": "This trick has been recently added to the official trick-list. Go out and try to land it!"
+        "description": "This trick has been recently added to the official trick list. Go out and try to land it!"
       }
     }
   },
